% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_kruskal.R
\name{internal_kruskal}
\alias{internal_kruskal}
\title{kruskal_groups internal helper function}
\usage{
internal_kruskal(df, y, x, pos, alpha, p_adj)
}
\arguments{
\item{df}{A data.frame.}

\item{y}{The name of the y (dependent) variable.}

\item{x}{The name of the x (independent) variable.}

\item{pos}{The y value at which the letter annotation should be placed in the plot.}

\item{alpha}{The significance threshhold alpha for the the Tukey-HSD test.}
}
\description{
Internal helper that is called after the grouping.
}
\keyword{internal}
