% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_stats.R
\name{internal_stats}
\alias{internal_stats}
\title{letter_groups internal helper function}
\usage{
internal_stats(
  df,
  y,
  x,
  stat_method,
  print_position,
  print_adjust,
  stat_alpha,
  p_adj_method
)
}
\arguments{
\item{df}{A data frame.}

\item{y}{The name of the y (dependent) variable.}

\item{x}{The name of the x (independent) variable.}

\item{stat_method}{The statistical test applied. Either "tukey" or "kruskal".}

\item{print_position}{The y position at which the letter annotation will be placed in the plot. One of "above", "mean", "below", or a numeric value.}

\item{print_adjust}{Adjustment of the letter position multiples of the overall standard deviations. Defaults to 1.}

\item{stat_alpha}{The significance threshold alpha. Defaults to 0.05.}

\item{p_adj_method}{Method for p value adjustment, see ?dunn.test::dunn.test for details. One of "none", "bonferroni", "sidak", "holm", "hs", "hochberg", "bh", "by". Defaults to "holm".}
}
\description{
Internal helper that is called after the grouping.
}
\keyword{internal}
