% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_tukey.R
\name{internal_tukey}
\alias{internal_tukey}
\title{tukey_groups internal helper function}
\usage{
internal_tukey(df, y, x, print_position, print_adjust, stat_alpha)
}
\arguments{
\item{df}{A data.frame.}

\item{print_position}{The y position at which the letter annotation will be placed in the plot. One of "above", "mean", "below", or a numeric value.}

\item{print_adjust}{Adjustment of the letter position multiples of the overall standard deviations. Defaults to 1.}

\item{stat_alpha}{The significance threshold alpha for the the Tukey HSD test. Defaults to 0.05.}

\item{y_var}{The name of the y (dependent) variable.}

\item{x_var}{The name of the x (independent) variable.}
}
\description{
Internal helper that is called after the grouping.
}
\keyword{internal}
