% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_groups.R
\name{kruskal_groups}
\alias{kruskal_groups}
\title{Kruskal-LSD plot annotation}
\usage{
kruskal_groups(df, y_var, x_var, pos_var, alpha_var, p_adj_var, ...)
}
\arguments{
\item{df}{A data.frame.}

\item{y_var}{The name of the y (dependent) variable.}

\item{x_var}{The name of the x (independent) variable.}

\item{pos_var}{The y value at which the letter annotation should be placed in the plot.}

\item{alpha_var}{The significance threshhold alpha for the the Tukey-HSD test.}

\item{p_adj_var}{Method for p value adjustment. One of "none","holm","hommel", "hochberg", "bonferroni", "BH", "BY" or "fdr".}

\item{...}{Any number of grouping variables for faceted plots.}
}
\description{
This is a simple helper function that outputs a data frame that can be used to annotate your plot with letter codes according to significant differences between groups. In addition to the data and formula, the function takes as input a positional variable for placement on the plot, an alpha for the Tukey-HSD test, and grouping variables for faceted plots.
}
\examples{
test_data <- tibble(value = rnorm(1000), sample = sample(LETTERS[1:5], 1000, replace = TRUE), group = sample(LETTERS[24:26], 1000, replace = TRUE))
annotation_df <- kruskal_groups(test_data, value, sample, 1, 0.001, group)
}
\keyword{Kruskal,}
\keyword{annotation}
