% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_groups.R
\name{kruskal_groups}
\alias{kruskal_groups}
\title{Kruskal-Wallis plot annotation}
\usage{
kruskal_groups(df, y, x, ..., print_position = 0, print_adjust = 1,
  stat_alpha = 0.05, p_adj_method = "holm")
}
\arguments{
\item{df}{A data.frame.}

\item{...}{Any number of grouping variables for faceted plots.}

\item{print_position}{The y position at which the letter annotation will be placed in the plot. One of "above", "mean", "below", or a numeric value.}

\item{print_adjust}{Adjustment of the letter position multiples of the overall standard deviations. Defaults to 1.}

\item{stat_alpha}{The significance threshold alpha for the the Kruskal-Wallis test. Defaults to 0.05.}

\item{p_adj_method}{Method for p value adjustment. One of "none","holm","hommel", "hochberg", "bonferroni", "BH", "BY" or "fdr". Defaults to "holm".}

\item{y_var}{The name of the y (dependent) variable.}

\item{x_var}{The name of the x (independent) variable.}
}
\description{
This is a simple wrapper function that outputs a data frame that can be used to annotate your plot with letter codes according to a Kruskal-Wallis test. In addition to the data and formula, the function takes as input a positional variable for placement on the plot, an alpha for the Kruskal-Wallis test, the method for p-value adjustment, and grouping variables for faceted plots. The underlying statistical functions are called from the 'agricolae' package.
}
\examples{
test_data <- tibble(value = rnorm(1000), sample = sample(LETTERS[1:5], 1000, replace = TRUE), group = sample(LETTERS[24:26], 1000, replace = TRUE))
annotation_df <- kruskal_groups(test_data, value, sample, group, stat_alpha = 0.001, print_position = "above", print_adjust = 0.5, p_adj_method = "bonferroni")
}
\keyword{Kruskal,}
\keyword{annotation}
\keyword{non-parametric,}
